backend/requirements.txt
# fastapi
# uvicorn
# pydantic
# pdfplumber
# pdf2image
# pillow
# pytesseract
# sentence-transformers
# faiss-cpu
# reportlab
# python-multipart
# # for local LLM
# transformers
# torch
# llama-cpp-python
# nltk
# requests
fastapi
uvicorn
requests
pydantic
faiss-cpu
sentence-transformers
